refactor(settings): decompose SettingsViewModel into state, prefs controller, and account delegate
Splits the 966-line SettingsViewModel into a 435-line orchestrator plus three
focused files. Pure refactor — public constructor and method surface unchanged,
so the DI graph and screen call sites are untouched.

- SettingsUiState.kt: the state class, SettingsCommand/LogsExportPayload/
  DEFAULT_COMMANDS, and the User.toDisplayData mapper.
- SettingsPreferencesController.kt: the DataStore read flows (intermediate
  holders, stateIn flows, the combine that merges them with imperative state
  into uiState) plus the matching write setters.
- AccountDelegate.kt: profile load (cancel-on-retry), avatar upload, balance,
  sign-out, and account deletion with the 403 to OTP step-up.
- Records the decomposition in MODULARIZATION.md.
